add page

+++ laundry/src/main/java/com/example/laundry/services/impl/PaymentServiceImpl.java
@@ -10,8 +10,6 @@ import com.example.laundry.services.PaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
-
 @Service
 public class PaymentServiceImpl implements PaymentService {
     @Autowired
@@ -30,8 +28,7 @@ public class PaymentServiceImpl implements PaymentService {
 
         return PaymentResponse.builder()
                 .orderId(Long.valueOf(order.getId().toString()))
-                // ép kiểu
-                .totalAmount(BigDecimal.valueOf(order.getTotalAmount()))
+                .totalAmount(order.getTotalAmount())
                 .accountHolder("Pham Minh Chi")
                 .bankName("MB Bank")
                 .accountNumber("0943869063")
